[automation] Rename gym script to workout
Having a gym script have a `--gym` argument just felt weird and wrong.
